tests: refactor duplicate code
- Include helpers.py for common functions
