+++ tests/bug_checks_functions.py
@@ -2,6 +2,7 @@ from scipy.signal import find_peaks
 import matplotlib.pyplot as plt
 import libm2k
 
+from helpers import get_result_files, plot_to_file
 from open_context import ctx_timeout, ctx, ain, aout, trig
 
 gen_reports = True
@@ -9,16 +10,8 @@ gen_reports = True
 
 def test_dac_artifact(ain, aout, trig, chn):
     # check for DAC artifacts at slow sample rates
-    if gen_reports:
-        from create_files import results_file, results_dir, csv, open_files_and_dirs
-        if results_file is None:
-            file, dir_name, csv_path = open_files_and_dirs()
-        else:
-            file = results_file
-            dir_name = results_dir
-            csv_path = csv
-    else:
-        file = []
+    file_name, dir_name, csv_path = get_result_files(gen_reports)
+
     buffer_size = 4000
     sampling_frequency_in = 1000000
     sampling_frequency_out = 750000
@@ -64,32 +57,3 @@ def generate_clock_signal():
     for i in range(256):
         buffer.append(0)
     return buffer
-
-
-def plot_to_file(title, data, dir_name, filename, xlabel=None, ylabel=None, data1=None):
-    # Saves the plots in a separate folder
-    # Arguments:
-    #    title  -- Title of the plot
-    #    data  -- Data to be plotted
-    #    filename  -- Name of the file with the plot
-    # Keyword Arguments:
-    #    xlabel  -- Label of x-Axis (default: {None})
-    #    ylabel  -- Label of y-Axis(default: {None})
-    #    data1  --  Data that should be plotted on the same plot(default: {None})
-    # plot the signals in a separate folder
-    plt.title(title)
-    if xlabel is not None:  # if xlabel and ylabel are not specified there will be default values
-        plt.xlabel(xlabel)
-    else:
-        plt.xlabel('Samples')
-    if ylabel is not None:
-        plt.ylabel(ylabel)
-    else:
-        plt.ylabel('Voltage [V]')
-    plt.grid(visible=True)
-    plt.plot(data)  # if a second set of data must be printed (for ch0 and ch1 phase difference in this case)
-    if data1 is not None:
-        plt.plot(data1)
-    plt.savefig(dir_name + "/" + filename)
-    plt.close()
-    return
